+++ cli/bootstrap/custom-app/src/App.js
@@ -1,0 +1,10 @@
+import React from 'react';
+import ReactTV from 'react-tv';
+
+class ReactTVApp extends React.Component {
+  render() {
+    return <div>{{REACTTVAPP}}</div>;
+  }
+}
+
+ReactTV.render(<ReactTVApp />, document.querySelector('#root'));