+++ cli/scripts/index.js
@@ -1,4 +1,3 @@
 module.exports = {
   runWebOS: require('./webos/run'),
-  runWebOSDev: require('./webos/run-dev'),
 };