+++ src/modules/WebOSSimulator.js
@@ -1,0 +1,16 @@
+/**
+ * Copyright (c) 2017-present, Raphael Amorim.
+ *
+ * This source code is licensed under the MIT license found in the
+ * LICENSE file in the root directory of this source tree.
+ *
+ * @flow
+ */
+
+import React from 'react';
+
+const WebOSSimulator = ({children}: {children?: any}) => (
+  <div class="react-tv-webos-simulator">{children}</div>
+);
+
+export default WebOSSimulator;