+++ CHANGELOG.md
@@ -1,3 +1,28 @@
+# 0.3.0-beta.0
+
+- stable CLI react-tv init
+- fixes on react-tv run-webos
+- on process.end remove/cleanup files
+- on virtualbox call remove/cleanup files
+
+# 0.3.0-alpha.2
+
+- CLI: `react-tv init`
+- CLI: `react-tv init <app-name>`
+- add tests for CLI behavior
+- add .npmignore
+- add `jest:ci` job
+
+# 0.3.0-alpha.1
+
+- migrate to react-reconciler 0.6.0
+- check if is need to format code on CI
+- add benchmark example
+- fix update render tor nodeText
+- update postinstall to prebublishOnly (newest npm)
+- remove: react-tv run-webos-dev
+- react-tv cli {init} breaking changes
+
 # 0.2.3
 
 - use ReactReconciler from `react-reconciler` instead from `react-dom`